re-organize imports for migrate sub-commands
